requirements.txt
gunicorn
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
attrs==24.2.0
bcrypt==4.2.0
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
dnspython==2.6.1
Faker==28.4.1
Flask==3.0.3
Flask-PyMongo==2.3.0
frozenlist==1.4.1
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.1.0
numpy==2.1.1
pandas==2.2.2
pillow==10.4.0
pycparser==2.22
pycryptodome==3.20.0
PyJWT==2.9.0
pymongo==4.8.0
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
qrcode==7.4.2
requests==2.32.3
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.4
yarl==1.11.1
yolov5